+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - ISelectable interface
 - ISubmitable interface
+- SelectableAudioMenu component
 
 ### Changed
 - DelayedButton implements ISelectable and ISubmitable interfaces